switch chain routing uint8 -> uint16

+++ contracts/OgSwap/interfaces/wormhole.sol
@@ -3,8 +3,8 @@ pragma solidity >=0.8.0;
 
 interface Wormhole {
     function lock(
-        uint8 chainType,
-        uint8 chainId,
+        uint16 chainType,
+        uint16 chainId,
         uint32 nonce,
         bytes calldata customPayload
     ) external;
